Revamp the leave command
Took 13 minutes
